This was a FINAL project
If you guys want to use a simple inventory management, you can check this one out. =)

This was made in Visual Basic 2010 and a database using SQLyog and XAMPP. It is our finals project and if you guys want to implement something new. Don't hesitate trying it out and credit.